Delete event now triggered on metadata and annotation delete
git-svn-id: https://code.elgg.org/elgg/trunk@3305 36083f99-b078-4883-b0ff-0f9b5a30f544

+++ engine/lib/annotations.php
@@ -785,8 +785,12 @@
 		$id = (int)$id;
 		
 		$access = get_access_sql_suffix();
+		$annotation = get_annotation($id);
 		
-		return delete_data("DELETE from {$CONFIG->dbprefix}annotations  where id=$id and $access");
+		if (trigger_elgg_event('delete', 'annotation', $annotation))
+			return delete_data("DELETE from {$CONFIG->dbprefix}annotations  where id=$id and $access");
+			
+		return false;
 	}
 	
 	/**
